Removing logging and unnecessary logic
After looking into more, I noticed that using jar:file CAN use C:, D:, E: etc as long as it preceded by a '/'. So instead of checking and removing for letter drives, just always add a slash in front if there isn't one. This needs to be there regardless of system, so just removed a bunch of garbage logic. Also forgot to remove logging so took that out.